# src/lumio-web/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Lumio API",
    "description": "Offline digitale nalatenschap beheer — Lumio",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/videoboodschappen/uploaden": {
      "post": {
        "tags": [
          "Videoboodschappen"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Bestand",
                  "Titel"
                ],
                "type": "object",
                "properties": {
                  "Bestand": {
                    "type": "string",
                    "format": "binary"
                  },
                  "Titel": {
                    "type": "string"
                  },
                  "Beschrijving": {
                    "type": "string"
                  },
                  "OntvangerIds": {
                    "type": "string"
                  },
                  "DuurSeconden": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              },
              "encoding": {
                "Bestand": {
                  "style": "form"
                },
                "Titel": {
                  "style": "form"
                },
                "Beschrijving": {
                  "style": "form"
                },
                "OntvangerIds": {
                  "style": "form"
                },
                "DuurSeconden": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VideoboodschapResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoboodschapResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoboodschapResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/videoboodschappen/{id}/stream": {
      "get": {
        "tags": [
          "Videoboodschappen"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/videoboodschappen/{id}": {
      "patch": {
        "tags": [
          "Videoboodschappen"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoboodschapUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoboodschapUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VideoboodschapUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VideoboodschapResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoboodschapResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoboodschapResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Videoboodschappen"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/videoboodschappen/preview": {
      "post": {
        "tags": [
          "Videoboodschappen"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Bestand"
                ],
                "type": "object",
                "properties": {
                  "Bestand": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "Bestand": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/videoboodschappen/preview/{tempId}/stream": {
      "get": {
        "tags": [
          "Videoboodschappen"
        ],
        "parameters": [
          {
            "name": "tempId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/videoboodschappen/preview/{tempId}": {
      "delete": {
        "tags": [
          "Videoboodschappen"
        ],
        "parameters": [
          {
            "name": "tempId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/videoboodschappen/voor-erfgenaam/{erfgenaamId}": {
      "get": {
        "tags": [
          "Videoboodschappen"
        ],
        "parameters": [
          {
            "name": "erfgenaamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoboodschapResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoboodschapResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoboodschapResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OntvangerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "erfgenaamId": {
            "type": "string",
            "format": "uuid"
          },
          "naam": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VideoboodschapResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "titel": {
            "type": "string",
            "nullable": true
          },
          "beschrijving": {
            "type": "string",
            "nullable": true
          },
          "bestandsNaam": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "bestandsGrootte": {
            "type": "integer",
            "format": "int64"
          },
          "duurSeconden": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ontvangers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OntvangerResponse"
            },
            "nullable": true
          },
          "aangemaaktOp": {
            "type": "string",
            "format": "date-time"
          },
          "gewijzigdOp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "VideoboodschapUpdateRequest": {
        "type": "object",
        "properties": {
          "titel": {
            "type": "string",
            "nullable": true
          },
          "beschrijving": {
            "type": "string",
            "nullable": true
          },
          "ontvangerIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}